WFLY-18264 Convert TimerAttributeDefinition to ObjectListAttributeDefinition
